+++ psdist/visualization/__init__.py
@@ -1,4 +1,4 @@
-import psdist.visualization.cloud
 import psdist.visualization.image
+import psdist.visualization.points
 from psdist.visualization.grid import *
 from psdist.visualization.visualization import *